A little refactoring. Converted field injectors to property injectors. Fixed bug in deleteTheme method

+++ src/main/java/repositories/CampaignRepository.java
@@ -16,32 +16,65 @@ import java.util.List;
 
 public class CampaignRepository {
 
-    @Inject
-    private Provider<EntityManager> entityManagerProvider;
 
+    private final Provider<EntityManager> entityManagerProvider;
+
+    @Inject
+    public CampaignRepository(Provider<EntityManager> entityManagerProvider) {
+        this.entityManagerProvider = entityManagerProvider;
+    }
+
+    /**
+     *
+     * @return
+     */
     @Transactional
     public List<Campaign> findAll() {
         Query q = getEntityManager().createNamedQuery("Campaign.findAll");
         return q.getResultList();
     }
 
+    /**
+     *
+     * @param campaignId
+     * @return
+     * @throws IllegalArgumentException
+     */
     @Transactional
-    public Campaign findCampaignById(Long campaignId) throws IllegalArgumentException {
+    public Campaign findCampaignById(Long campaignId) {
         return getEntityManager().find(Campaign.class, campaignId);
     }
 
+    /**
+     *
+     * @param campaign
+     * @throws CampaignException
+     */
     @Transactional
     public void save(Campaign campaign) throws CampaignException {
         getEntityManager().persist(campaign);
     }
 
+    /**
+     *
+     * @param campaign
+     */
     @Transactional
     public void update(Campaign campaign) {
         getEntityManager().persist(campaign);
     }
 
+    /**
+     *
+     * @param campaignId
+     */
     @Transactional
-    public void deleteCampaign(Campaign campaign) {
+    public void deleteCampaign(Long campaignId) throws CampaignException {
+        Campaign campaign = findCampaignById(campaignId);
+
+        if (campaign == null) {
+            throw new CampaignException("The campaign you're trying to update does not exist.");
+        }
         getEntityManager().remove(campaign);
     }
